Rename all folders and imports to use 'common' instead of 'commons'
