+++ src/common/i18n/locales/index.js
