+++ src/common/resources/api/example-api.js
