+++ src/common/resources/example.js
